taller-maestro-backend/auth_user/serializers.py
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from django.contrib.auth import authenticate
from rest_framework import serializers
from django.contrib.auth import get_user_model

User = get_user_model()

class EmailTokenObtainPairSerializer(TokenObtainPairSerializer):
    username_field = 'email'

    def validate(self, attrs):
        credentials = {
            self.username_field: attrs.get(self.username_field),
            'password': attrs.get('password')
        }
        user = authenticate(**credentials)
        if user is None:
            raise serializers.ValidationError('No se pudo autenticar con las credenciales proporcionadas')
        data = super().validate(attrs)
        # Añadir datos extra del usuario al token
        data['first_name'] = user.first_name
        data['last_name'] = user.last_name
        data['email'] = user.email
        return data
    
