taller-maestro-backend/requirements.txt
Django>=5.2
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
pillow==11.2.1
sqlparse==0.5.3
